requirements.txt
#
# EmBCI Python module requirements.txt.
#
# Automatically generated file.
#
# DO NOT EDIT THIS FILE -- changes will be overwritten.
#


# [testing]
pytest

# [network]
beaker
bottle
bottle-cork
gevent
gevent-websocket

# [machine-learning]
keras
sklearn

# [drivers]
gpio4
pyserial
spidev

# [data]
mne
numpy
pylsl
scipy

# [necessary]
decorator
dill
gym
pillow
pyhht
pywavelets
pyzmq
six


